cinema-11/base_requirements.txt
Flask==1.0.0
pytest
requests
opentelemetry-sdk==1.0.0rc1
opentelemetry-api==1.0.0rc1
opentelemetry-instrumentation==0.18b0
opentelemetry-exporter-jaeger==1.0.0rc1
opentelemetry-instrumentation-flask==0.18b1
opentelemetry-instrumentation-requests==0.18b1
docker
kubernetes
circuitbreaker